otorisasi
konfigurasi RBAC dengan CASL
memproteksi api produk
memproteksi api categories
memproteksi api tags

+++ app/categories/controller.js
@@ -1,7 +1,22 @@
 const Category = require('./model')
 
+// import policyfor casl
+const { policyFor } = require('../policy')
+
 async function store(req, res, next) {
     try {
+
+        //--- cek policy ---/
+        let policy = policyFor(req.user)
+
+        if (!policy.can('create', 'Category')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses untuk menambah kategori'
+            })
+        }
+        //-----------------//
+
         // (1) tangkap payload dari _client request_ 
         let payload = req.body
 
@@ -32,6 +47,18 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+
+        //--- cek policy ---/
+        let policy = policyFor(req.user)
+
+        if (!policy.can('update', 'Category')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses untuk mengupdate kategori'
+            })
+        }
+        //-----------------//
+
         // (1) tangkap payload dari _client request_ 
         let payload = req.body
 
@@ -59,6 +86,18 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
     try {
+
+        //--- cek policy ---/
+        let policy = policyFor(req.user)
+
+        if (!policy.can('delete', 'Category')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses untuk menghapus kategori'
+            })
+        }
+        //-----------------//
+
         // (1) cari dan hapus categori di MongoDB berdasarkan field _id
         let deleted = await Category.findByIdAndDelete({ _id: req.params.id })
